student-course-allocation/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
mysql-connector-python==8.2.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==41.0.7
typing-extensions==4.9.0
gunicorn==21.2.0
